Доработан механизм добавления игроков в базу

+++ backend/src/main/java/kernel/services/StatsSiteServices.java
@@ -184,7 +184,7 @@ public class StatsSiteServices
 
                 for(ServerName srv: srvNames)
                 {
-                    if(srv.getId() == game.getServerNameId())
+                    if((int)srv.getId() == (int)game.getServerNameId())
                     {
                         gameStruct.setServerName(srv.getName());
                         break;
